requirements.txt
httpx==0.27.2
beautifulsoup4==4.12.3
lxml==5.3.0
Pillow==11.0.0
typer==0.12.5
python-dotenv==1.0.1
